+++ src/app/app.module.ts
@@ -20,18 +20,18 @@ import 'reflect-metadata';
 import '../polyfills';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { BaseRemoteProxy } from './common/base-remote-proxy';
-import { ClipboardWatcher } from './common/clipboard-watcher';
+import { GitHubApi } from './common/api/github-api';
+import { ProductDetails } from './common/application/product-details';
 import { Delayer } from './common/delayer';
-import { Desktop } from './common/desktop';
-import { DocumentProxy } from './common/document-proxy';
-import { ElectronRemoteProxy } from './common/electron-remote-proxy';
-import { Environment } from './common/environment';
-import { FileSystem } from './common/file-system';
-import { GitHubApi } from './common/github-api';
+import { BaseRemoteProxy } from './common/io/base-remote-proxy';
+import { ClipboardWatcher } from './common/io/clipboard-watcher';
+import { Desktop } from './common/io/desktop';
+import { DocumentProxy } from './common/io/document-proxy';
+import { ElectronRemoteProxy } from './common/io/electron-remote-proxy';
+import { Environment } from './common/io/environment';
+import { FileSystem } from './common/io/file-system';
+import { RemoteProxy } from './common/io/remote-proxy';
 import { Logger } from './common/logger';
-import { ProductDetails } from './common/product-details';
-import { RemoteProxy } from './common/remote-proxy';
 import { BaseSettings } from './common/settings/base-settings';
 import { Settings } from './common/settings/settings';
 import { AboutComponent } from './components/about/about.component';
