+++ src/app/common/application/application-paths.spec.ts
