+++ src/app/common/application/application-paths.ts
