+++ src/app/common/application/product-details.spec.ts
