--- src/app/common/color-theme.ts
@@ -1,4 +1,0 @@
-export class ColorTheme {
-    constructor(public name: string, public displayName: string, public color: string) {
-    }
-}
