--- src/app/common/events.js
@@ -1,10 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var Events = /** @class */ (function () {
-    function Events() {
-    }
-    Events.windowFocusChangedEvent = '59e9967d-8f0d-4b01-b48d-efecdd95c986';
-    return Events;
-}());
-exports.Events = Events;
-//# sourceMappingURL=events.js.map
