--- src/app/common/events.ts
@@ -1,3 +1,0 @@
-export class Events {
-    public static readonly windowFocusChangedEvent: string = '59e9967d-8f0d-4b01-b48d-efecdd95c986';
-}
