+++ src/app/common/io/clipboard-watcher.mock.ts
