+++ src/app/common/io/desktop.ts
