--- src/app/common/utils.ts
@@ -1,5 +1,0 @@
-export class Utils {
-    public static async sleep(milliseconds: number): Promise<void> {
-        return new Promise(resolve => setTimeout(resolve, milliseconds));
-    }
-}
