+++ src/app/components/dialogs/error-dialog/error-dialog.component.ts
@@ -1,8 +1,8 @@
 import { Component, Inject, OnInit, ViewEncapsulation } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
-import { Constants } from '../../../common/constants';
-import { Desktop } from '../../../common/desktop';
-import { FileSystem } from '../../../common/file-system';
+import { Constants } from '../../../common/application/constants';
+import { Desktop } from '../../../common/io/desktop';
+import { FileSystem } from '../../../common/io/file-system';
 @Component({
     selector: 'app-error-dialog',
     templateUrl: './error-dialog.component.html',
