+++ src/app/services/appearance/appearance.service.spec.ts
@@ -1,17 +1,17 @@
 import { OverlayContainer } from '@angular/cdk/overlay';
 import { Observable, Subject } from 'rxjs';
 import { IMock, Mock, Times } from 'typemoq';
-import { BaseRemoteProxy } from '../../common/base-remote-proxy';
-import { Constants } from '../../common/constants';
-import { Desktop } from '../../common/desktop';
-import { DocumentProxy } from '../../common/document-proxy';
-import { FileSystem } from '../../common/file-system';
-import { FontSize } from '../../common/font-size';
+import { Constants } from '../../common/application/constants';
+import { BaseRemoteProxy } from '../../common/io/base-remote-proxy';
+import { Desktop } from '../../common/io/desktop';
+import { DocumentProxy } from '../../common/io/document-proxy';
+import { FileSystem } from '../../common/io/file-system';
 import { Logger } from '../../common/logger';
 import { BaseSettings } from '../../common/settings/base-settings';
 import { AppearanceService } from './appearance.service';
 import { BaseAppearanceService } from './base-appearance.service';
 import { DefaultThemesCreator } from './default-themes-creator';
+import { FontSize } from './theme/font-size';
 import { Theme } from './theme/theme';
 import { ThemeCoreColors } from './theme/theme-core-colors';
 import { ThemeCreator } from './theme/theme-creator';
