+++ src/app/services/convert/convert.service.spec.ts
@@ -3,8 +3,8 @@ import * as path from 'path';
 import { Subscription } from 'rxjs';
 import { IMock, It, Mock, Times } from 'typemoq';
 import { AudioFormat } from '../../common/audio-format';
-import { Environment } from '../../common/environment';
-import { FileSystem } from '../../common/file-system';
+import { Environment } from '../../common/io/environment';
+import { FileSystem } from '../../common/io/file-system';
 import { Logger } from '../../common/logger';
 import { BaseSettings } from '../../common/settings/base-settings';
 import { BaseConvertService } from './base-convert.service';
