+++ src/app/services/convert/dependency-checker-factory.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { FileSystem } from '../../common/file-system';
+import { FileSystem } from '../../common/io/file-system';
 import { Logger } from '../../common/logger';
 import { DependencyChecker } from './dependency-checker';
 import { YoutubeDownloaderConstants } from './youtube-downloader-constants';
