+++ src/app/services/convert/dependency-checker.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import * as path from 'path';
-import { FileSystem } from '../../common/file-system';
+import { FileSystem } from '../../common/io/file-system';
 import { Logger } from '../../common/logger';
 import { Strings } from '../../common/strings';
