+++ src/app/services/convert/video-converter.factory.spec.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert';
 import { Mock } from 'typemoq';
-import { Environment } from '../../common/environment';
+import { Environment } from '../../common/io/environment';
 import { Logger } from '../../common/logger';
 import { VideoConverterFactory } from './video-converter.factory';
 import { YoutubeVideoConverter } from './youtube-video-converter';
