+++ src/app/services/convert/youtube-video-converter.ts
@@ -1,6 +1,6 @@
 import * as child from 'child_process';
 import { AudioFormat } from '../../common/audio-format';
-import { Environment } from '../../common/environment';
+import { Environment } from '../../common/io/environment';
 import { Logger } from '../../common/logger';
 import { Strings } from '../../common/strings';
 import { ConversionResult } from './conversion-result';
